Include see also pointer to tidyr::extract
I turned to str_match and str_extract for doing what tidyr::extract does (and tidy::extract is a much more convenient form!). So I added a cross-link to the see also.

Thanks to markdly at community.rstudio.org showing me the right way! https://community.rstudio.com/t/getting-parts-of-string-using-regex-and-capturing-parens-in-mutate