Fix typo in regular expressions vignette
Fixed typo that `[\^]` matches `^` and not `\`. Also changed order to be consistent with presentation in regexp.

Is it worth mentioning here that no backslash is needed if `-` is the first character or if `^` is not the first character in the class? In other words `[\^\-]` is equivalent to `[-^]`. (I guess not, it is easier to just learn the rules as escape both inside the `[]`.)