Fix str_view_all() deprecation warning
In 1.5.0, str_view() was improved and as a result str_view_all() [was made redundant and was deprecated.](https://www.tidyverse.org/blog/2022/12/stringr-1-5-0/#str_view)  However, the deprecation warning that is displayed when str_view_all() is called states that str_view() is deprecated and that str_view_all() should be used. This change corrects the warning to say that str_view_all() is deprecated and that str_view() should be used. Fixes issue #488.